% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_addMotifAnnotation.R
\name{addMotifAnnotation}
\alias{addMotifAnnotation}
\title{to do}
\usage{
addMotifAnnotation(AUCellOutput, nesThreshold = 3, digits = 3,
  motifAnnot_direct = NULL, motifAnnot_indirect = NULL,
  highlightTFs = NULL)
}
\arguments{
\item{to}{do}
}
\value{
to do Columna TF
}
\description{
to do
}

