% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_dbReggionsLoc.R
\docType{data}
\name{dbRegionsLoc}
\alias{dbRegionsLoc}
\alias{dbRegionsLoc_hg19}
\alias{dbRegionsLoc_mm9}
\title{Genomic location for the database regions}
\description{

}
\details{
Documentation for the data
}
\seealso{
https://resources.aertslab.org/cistarget
}
\keyword{datasets}
