% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getSignificantGenes_v2.2.R
\name{getSignificantGenes}
\alias{getSignificantGenes}
\title{to do}
\usage{
getSignificantGenes(geneSet, rankings, signifRankingNames = NULL,
  method = "iCisTarget", maxRank = 5000, plotCurve = FALSE,
  genesFormat = c("geneList", "incidMatrix"), nCores = 1, digits = 3)
}
\arguments{
\item{to}{do}
}
\value{
to do
}
\description{
to do
}
\examples{
# to do
}

