% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/00_importRankings.R
\name{importRankings}
\alias{importRankings}
\title{Import the motif databases for RcisTarget.}
\usage{
importRankings(dbFile, columns = NULL, dbDescr = NULL)
}
\arguments{
\item{dbFile}{.feather file containing the rankings}

\item{columns}{Columns to import from the .feather file 
(e.g. only selected genes or regions)}

\item{dbDescr}{The description fields are typically imported from the
".descr" file with the same filename as the \code{dbFile}. 
Otherwise they can be manually set e.g.: 

\code{dbDescr=list(colType="gene", rowType="motif", 
org="Human", genome="hg19", maxRank=Inf, description="")}}
}
\value{
rankingRcisTarget object with the following slots:
#' \itemize{
\item rankings: data.frame containing the rankings
\item colType: 'gene'or 'region'
\item rowType: 'motif' or the type of feature is stored (e.g. ChipSeq)
\item org: human/mouse/fly
\item genome: hg19, mm9, ...
\item description: global description, summary, or any other information
\item maxRank: Maximum ranking included in the database, 
higher values are converted to Inf.
}
}
\description{
The rankings are typically loaded from a .feather file 
with \code{importRankings()}.

If the associated .descr file is available, 
it will also load the description of the database.
}
\examples{
## Loading from a .feather file (the .descr file is read automatically):
#motifRankings<-importRankings("hg19-500bp-upstream-7species.mc9nr.feather")

## The annotations for Motif collection 9 (sufix 'mc9nr') 
# are already included in RcisTarget, and can be loaded with:
data(motifAnnotations_hgnc)

## For other versions, import the appropiate annotation. e.g.:
# annotDb <- importAnnotations("motifs-v9-nr.hgnc-m0.001-o0.0.tbl")
# optional: motifsInRanking <- getRanking(motifRankings)$features
}
