% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_motifsAUC.R
\docType{class}
\name{motifsAUC-class}
\alias{motifsAUC-class}
\alias{motifsAUC}
\title{Class to store AUC}
\description{
Class with only one slot: @AUC, a matrix which contains the AUC for the gene-sets & motifs.
}
