% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_addMotifAnnotation.R
\name{addMotifAnnotation}
\alias{addMotifAnnotation}
\title{Add motif annotation}
\usage{
addMotifAnnotation(auc, nesThreshold = 3, digits = 3,
  motifAnnot_direct = NULL, motifAnnot_inferred = NULL,
  highlightTFs = NULL)
}
\arguments{
\item{auc}{Output from calcAUC.}

\item{nesThreshold}{Numeric. NES threshold to calculate the motif significant (3.0 by default). The NES is calculated -for each motif- based on the AUC distribution of all the motifs for the gene-set [(x-mean)/sd].}

\item{digits}{Integer. Number of digits for the AUC and NES in the output table.}

\item{motifAnnot_direct}{Motif annotation database containing DIRECT annotations of the motif to transcription factors. The names should match the ranking column names.}

\item{motifAnnot_inferred}{Motif annotation database containing the inferred annotations of the motif to transcription factors based on motif similarity.}

\item{highlightTFs}{Character. If a list of transcription factors is provided,
the column TFinDB in the otuput table will indicate whether any of those TFs are included within the 'direct' annotation (two asterisks, **)
or 'inferred' annotation (one asterisk, *) of the motif.
The vector can be named to indicate which TF to highlight for each gene-set. Otherwise, all TFs will be used for all geneSets.}
}
\value{
\code{\link[data.table]{data.table}} with the folowing columns:
\itemize{
\item geneSet: Name of the gene set
\item motif: ID of the motif (colnames of the ranking, it might be other kind of feature)
\item NES: Normalized enrichment score of the motif in the gene-set
\item AUC: Area Under the Curve (used to calculate the NES)
\item TFinDB: Indicates whether the highlightedTFs are included within the direct annotation (two asterisks, **) or inferred annotation (one asterisk, *)
\item TF_direct: Transcription factors annotated to the motif according to 'direct annotation'.
\item TF_inferred: Transcription factors annotated to the motif according to 'inferred annotation'.
}
}
\description{
Select significant motifs and/or annotate motifs to transcription factors.
The motifs are considered significantly enriched if they pass the the Normalized Enrichment Score (NES) threshold.
}
\examples{

##################################################
# Setup & previous steps in the workflow:
# Gene sets
txtFile <- paste(file.path(system.file('examples', package='RcisTarget')),
                 "hypoxiaGeneSet.txt", sep="/")
geneLists <- list(hypoxia=read.table(txtFile)[,1])

# Motif databases
# (Select the package/database according to the organism and distance around TSS)
library(RcisTarget.hg19.motifDatabases.20k)
data(hg19_10kbpAroundTss_motifRanking)
motifRankings <- hg19_10kbpAroundTss_motifRanking

# Fake-database with 5000 random motifs (to run the example faster)
# DO NOT use in real analyses!
set.seed(123)
motifRankings <- subset(hg19_10kbpAroundTss_motifRanking,
    sample(2:ncol(hg19_10kbpAroundTss_motifRanking@rankings), 5000), select="col")
motifRankings

# RcisTarget
# Step 1. Calculate AUC
motifs_AUC <- calcAUC(geneLists, motifRankings)

##################################################
# (This step: Step 2)
# Select significant motifs, add TF annotation & format as table
data(hg19_direct_motifAnnotation)

motifEnrichmentTable <- addMotifAnnotation(motifs_AUC,
   motifAnnot_direct=hg19_direct_motifAnnotation)

# Adding indirect annotation and modifying some options
data(hg19_bySimilarity_motifAnnotation)
motifEnrichment_wIndirect <- addMotifAnnotation(motifs_AUC, nesThreshold=2,
                        motifAnnot_direct=hg19_direct_motifAnnotation,
                        motifAnnot_bySimilarity=hg19_bySimilarity_motifAnnotation,
                        highlightTFs = "HIF1A", digits=3)

# Exploring the output:
# Note: Using the fake-database, these results are not meaningful.

# Number of enriched motifs (Over the given NES threshold)
nrow(motifEnrichmentTable)

# Interactive exploration
motifEnrichmentTable <- addLogo(motifEnrichmentTable)
DT::datatable(motifEnrichmentTable, filter="top", escape=FALSE, options=list(pageLength=50))
# Note: If using the fake database, the results of this analysis are meaningless

# The object returned is a data.table (for faster computation),
# which has a diferent syntax from the standard data.frame or matrix
# Feel free to convert it to a data.frame (as.data.frame())
motifEnrichmentTable[,1:6]


##################################################
# Next step (step 3, optional):
motifEnrichmentTable_wGenes <- addSignificantGenes(motifEnrichmentTable,
                                   geneSets=geneLists,
                                   rankings=motifRankings,
                                   method="aprox")

}
\seealso{
Next step in the workflow: \code{\link{addSignificantGenes}}.

Previous step in the workflow: \code{\link{calcAUC}}.

See the package vignette for examples and more details: \code{vignette("RcisTarget")}
}

