% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_getSignificantGenes_v2.2.R
\name{addSignificantGenes}
\alias{addSignificantGenes}
\title{to do}
\usage{
addSignificantGenes(aucTable, geneSets, rankings, maxRank = 5000,
  plotCurve = FALSE, genesFormat = "geneList", method = "aprox",
  nCores = 1)
}
\arguments{
\item{to}{do}
}
\value{
to do
}
\description{
to do
}
\details{
# required?
}

