% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_aucScores.R
\docType{class}
\name{aucScores-class}
\alias{aucScores-class}
\alias{aucScores}
\alias{getAUC}
\alias{show,aucScores-method}
\alias{getAUC}
\alias{getAUC,aucScores-method}
\title{Class to store the AUC scores for RcisTarget.}
\usage{
\S4method{show}{aucScores}(object)

getAUC(object)
}
\arguments{
\item{object}{Results from \code{calcAUC}.}
}
\value{
\itemize{
\item show: Prints a summary of the object
\item getAUC: Returns the matrix containing the AUC scores
}
}
\description{
Contains the AUC scores for each gene- or region-set.
They can be accessed through \code{getAUC()} and the regular methods
(i.e. nrow, rownames... ) available for SummarizedExperiment objects.
}
\examples{
############# Fake run of RcisTarget ########
set.seed(123)
motifRankings <- fakeDatabase()
motifRankings

geneLists <- list(geneSet=sample(getRanking(motifRankings)$rn, 100))

motifs_AUC <- calcAUC(geneLists, motifRankings)
##############################################

#Exploring the output:
motifs_AUC

class(motifs_AUC)

# Extracting the AUC matrix:
getAUC(motifs_AUC)[,1:5]

# Subsetting and regular manipulation methods are also available:
motifs_AUC[1,]
motifs_AUC[,3:4]

dim(motifs_AUC)
nrow(motifs_AUC)
ncol(motifs_AUC)
colnames(motifs_AUC)
rownames(motifs_AUC)
}
