% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_fakeDatabase.R
\name{fakeDatabase}
\alias{fakeDatabase}
\title{Creates a small fake database to use in examples}
\usage{
fakeDatabase(nGenes = 1000, nMotifs = 500, incAnnotation = FALSE)
}
\arguments{
\item{nGenes}{Number of genes to include in the ranking}

\item{nMotifs}{Number of motifs to include in the ranking}

\item{incAnnotation}{Whether to also generate a fake motif annotation}
}
\value{
Returns a small fake database, with random content but the right structure,
to run the examples.
}
\description{
Creates a small fake database to use in examples
}
\examples{
fakeDatabase()
}
