% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rankingRcisTarget.R
\docType{class}
\name{rankingRcisTarget-class}
\alias{rankingRcisTarget-class}
\alias{rankingRcisTarget}
\alias{rankingRcisTarget-class,}
\alias{rankingRcisTarget,}
\alias{getRanking,}
\alias{ncol,}
\alias{nrow,}
\alias{show,}
\alias{subset}
\alias{show,rankingRcisTarget-method}
\alias{getRanking}
\alias{getRanking,rankingRcisTarget-method}
\alias{getMaxRank}
\alias{getMaxRank,rankingRcisTarget-method}
\title{Class to store the motif databases for RcisTarget.}
\usage{
\S4method{show}{rankingRcisTarget}(object)

getRanking(x)

getMaxRank(x)
}
\arguments{
\item{object}{[method: show] rankingRcisTarget object to show}

\item{x}{[several methods] rankingRcisTarget object to apply the method}
}
\value{
\itemize{
\item show: Prints a summary of the object
\item getRanking: Returns the data.frame containing the rankings
\item ncol, nrow, subset: Returns the number of columns, rows
or a subset of the ranking
}
}
\description{
This class is only meant as container for the motif rankings
(for internal use). Modify content at your own risk.

Slots:
\itemize{
\item rankings: data.table with the rankings
\item colType: 'motif' or whatever other feature is stored (e.g. ChipSeq)
\item rowType: 'gene'or 'region'
\item org: human/mouse/fly
\item genome: hg19, mm9, ...
\item description: summary or any other information
}
}
\examples{
library(RcisTarget.hg19.motifDBs.cisbpOnly.500bp)
data("hg19_500bpUpstream_motifRanking_cispbOnly")
hg19_500bpUpstream_motifRanking_cispbOnly
class(hg19_500bpUpstream_motifRanking_cispbOnly)
}
