% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class_rankingWrapper.R
\docType{class}
\name{rankingWrapper-class}
\alias{getRanking}
\alias{rankingWrapper}
\alias{rankingWrapper-class}
\title{Class to store the motif databases for RcisTarget.}
\description{
This class is only meant for internal use. Modify at your own risk.
}

