% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/10_cisTarget.R
\name{cisTarget}
\alias{cisTarget}
\title{cisTarget}
\usage{
cisTarget(geneSets, motifRankings, motifAnnot_direct = NULL,
  motifAnnot_inferred = NULL, highlightTFs = NULL, nesThreshold = 3,
  aucMaxRank = 0.05 * nrow(motifRankings@rankings), geneErnMethod = "aprox",
  geneErnMmaxRank = 5000, nCores = 1, verbose = TRUE)
}
\arguments{
\item{geneSets}{List of gene-sets to analyze.
The gene-sets should be provided as \code{\link[GSEABase]{GeneSet}},
\code{\link[GSEABase]{GeneSetCollection}} or character list (see examples).}

\item{motifRankings}{Database of the appropiate organism and search-space
(i.e. 10kbp around- or 500bp upstream the TSS).
These objects are provided in separate packages:
\itemize{
\item \url{http://scenic.aertslab.org/downloads/databases/RcisTarget.mm9.motifDBs.20k_0.2.0.tar.gz}[RcisTarget.mm9.motifDatabases.20k_0.1.1.tar.gz] (Mouse)
\item \url{http://scenic.aertslab.org/downloads/databases/RcisTarget.hg19.motifDBs.20k_0.2.0.tar.gz}[RcisTarget.hg19.motifDatabases.20k_0.1.1.tar.gz] (Human)
\item -contact us- (Fly)
}
See the help files for more information:
i.e. \code{help(RcisTarget.hg19.motifDatabases)}}

\item{motifAnnot_direct}{Motif annotation database containing DIRECT
annotations of the motif to transcription factors.}

\item{motifAnnot_inferred}{Motif annotation database containing the expanded
annotations of the motif to transcription factors based on inference by
motif similarity.}

\item{highlightTFs}{Character.
If a list of transcription factors is provided, the column TFinDB in the
otuput table will indicate whether any of those TFs are included within the
direct annotation (two asterisks, **) or bySimilarity annotation
(one asterisk, *) of the motif. The vector can be named to indicate which
TF to highlight for each gene-set. Otherwise, all TFs will be used for all
geneSets.}

\item{nesThreshold}{Numeric.
NES threshold to calculate the motif significant (3.0 by default).
The NES is calculated -for each motif- based on the AUC distribution of
all the motifs for the gene-set [(x-mean)/sd].
The motifs are considered significantly enriched if they pass the the
 Normalized Enrichment Score (NES) threshold.}

\item{aucMaxRank}{Threshold to calculate the AUC.
In a simplified way, the AUC value represents the fraction of genes -within
 the top X genes in the ranking- that are included in the signature.
The parameter 'aucThresholdPERC' allows to modify the percentage of genes
(of the top of the ranking) that is used to perform this computation.
By default it is set to 5\% of the total number of genes in the rankings.
Common values range from 1 to 10\%.}

\item{geneErnMethod}{"iCisTarget" or "aprox". Method to identify the highly
ranked genes (see \link{addSignificantGenes} for details).}

\item{geneErnMmaxRank}{Maximum rank to take into account for the gene
enrichment recovery curve (see \link{addSignificantGenes} for details).}

\item{nCores}{Number of cores to use for computation.
Note: In general, using a higher number of cores (e.g. processes) decreases
overall running time. However, it also deppends on the available memory and
overall system load. Setting nCores too high might also decrease performance.}

\item{verbose}{Should the function show progress messages? (TRUE / FALSE)}
}
\value{
\code{\link[data.table]{data.table}} containing the over-represented
motifs (according to the selected NES threshold),
their statistics, annotation to transcription factors and the genes with
high enrichment of the motif.
}
\description{
Identifies DNA motifs significantly over-represented in a
gene-set.

This is the main function to run RcisTarget.
It includes on the following steps:
\itemize{
\item 1. Motif enrichment analysis (\link{calcAUC})
\item 2. Motif-TF annotation (\link{addMotifAnnotation})
\item 3. Selection of significant genes (\link{addSignificantGenes})
}
}
\examples{

# Example for running RcisTarget using cisTarget() function (workflow wrapper)


# A quick example can be run using a (small) fake-database & gene-set:
set.seed(123)
dbs <- fakeDatabase(incAnnotation=TRUE)
motifRankings <- dbs$ranking
motifAnnot_direct <- dbs$annotation
geneLists <- sample(rownames(motifRankings), 100)

# This is an example of code for a real run:
##################################################
\dontrun{
  #### Gene sets
  # As example, the package includes an Hypoxia gene set:
  txtFile <- paste(file.path(system.file('examples', package='RcisTarget')),
                   "hypoxiaGeneSet.txt", sep="/")
  geneLists <- list(hypoxia=read.table(txtFile)[,1])

  #### Databases
  # Select the package according to the organism and distance around TSS
  library(RcisTarget.hg19.motifDBs.20k)
  data(hg19_10kbpAroundTss_motifRanking)
  motifRankings <- hg19_10kbpAroundTss_motifRanking
  data(hg19_direct_motifAnnotation)
  motifAnnotation <- hg19_direct_motifAnnotation

  # Run (R)cisTarget
  motifEnrichmentTable_wGenes <- cisTarget(geneLists, motifRankings,
   motifAnnot_direct=hg19_direct_motifAnnotation,
   nesThreshold=3.5, geneErnMethod="aprox", nCores=2)
}

# Load results from analysis
load(paste(file.path(system.file('examples', package='RcisTarget')),
           "motifEnrichmentTable_wGenes.RData", sep="/"))
##################################################


# Exploring the output:
# Note: If using the fake-database, the results are not meaningful

# Number of enriched motifs (Over the given NES threshold)
nrow(motifEnrichmentTable_wGenes)

# Available info (columns)
colnames(motifEnrichmentTable_wGenes)

# The object returned is a data.table (for faster computation),
# which has a diferent syntax from the standard data.frame or matrix
# Feel free to convert it to a data.frame (as.data.frame())
class(motifEnrichmentTable_wGenes)
motifEnrichmentTable_wGenes[,1:5]

# Enriched genes
enrGenes <- as.character(motifEnrichmentTable_wGenes[1,"enrichedGenes"])
strsplit(enrGenes, ";")


# Interactive exploration
motifEnrichmentTable_wGenes <- addLogo(motifEnrichmentTable_wGenes)
DT::datatable(motifEnrichmentTable_wGenes[,1:9], escape = FALSE, filter="top",
              options=list(pageLength=5))
# Note: If using the fake database, the results of this analysis are meaningless
}
\seealso{
See the package vignette for examples and more details:
\code{vignette("RcisTarget")}
}
